changed filename to snake_case moved to src changed print(self)
